Split voice router into focused modules
