+++ frontend/src/components/ChapterEditor.test.tsx
@@ -627,87 +627,6 @@ describe('ChapterEditor', () => {
     expect(screen.queryByText('Finalizing')).not.toBeInTheDocument();
   });
 
-  it('keeps the queue button disabled while the header still shows queue status', () => {
-    const { rerender } = render(
-      <ChapterHeader
-        chapter={mockChapter as any}
-        title={mockChapter.title}
-        setTitle={vi.fn()}
-        saving={false}
-        hasUnsavedChanges={false}
-        onBack={vi.fn()}
-        selectedVoice=""
-        onVoiceChange={vi.fn()}
-        availableVoices={[]}
-        submitting={false}
-        queueLocked={false}
-        queuePending={false}
-        job={{ id: 'job-1', engine: 'mixed', status: 'running', progress: 1 } as any}
-        generatingSegmentIdsCount={0}
-        queueLabel="Complete"
-        queueTitle="Complete Chapter Audio"
-        onQueue={vi.fn()}
-        onStopAll={vi.fn()}
-      />
-    );
-
-    expect(screen.getByTitle('Already processing')).toBeDisabled();
-
-    rerender(
-      <ChapterHeader
-        chapter={mockChapter as any}
-        title={mockChapter.title}
-        setTitle={vi.fn()}
-        saving={false}
-        hasUnsavedChanges={false}
-        onBack={vi.fn()}
-        selectedVoice=""
-        onVoiceChange={vi.fn()}
-        availableVoices={[]}
-        submitting={false}
-        queueLocked={false}
-        queuePending={false}
-        job={{ id: 'job-1', engine: 'mixed', status: 'done', finished_at: Date.now() / 1000, progress: 1 } as any}
-        generatingSegmentIdsCount={0}
-        queueLabel="Complete"
-        queueTitle="Complete Chapter Audio"
-        onQueue={vi.fn()}
-        onStopAll={vi.fn()}
-      />
-    );
-
-    expect(screen.getByTitle('Already processing')).toBeDisabled();
-  });
-
-  it('shows working header state for active segment generation without a chapter render job', () => {
-    render(
-      <ChapterHeader
-        chapter={mockChapter as any}
-        title={mockChapter.title}
-        setTitle={vi.fn()}
-        saving={false}
-        hasUnsavedChanges={false}
-        onBack={vi.fn()}
-        selectedVoice=""
-        onVoiceChange={vi.fn()}
-        availableVoices={[]}
-        submitting={false}
-        queueLocked={false}
-        queuePending={false}
-        job={undefined}
-        generatingJob={{ id: 'job-seg', engine: 'mixed', status: 'running', progress: 0.4, started_at: Date.now() / 1000, eta_seconds: 9 } as any}
-        generatingSegmentIdsCount={2}
-        queueLabel="Complete"
-        queueTitle="Complete Chapter Audio"
-        onQueue={vi.fn()}
-        onStopAll={vi.fn()}
-      />
-    );
-
-    expect(screen.getByText('Processing')).toBeInTheDocument();
-    expect(screen.getByTitle('Already processing')).toBeDisabled();
-    expect(screen.getByText('40%')).toBeInTheDocument();
-  });
 
   it('ignores duplicate generate clicks while the same segments are already pending', async () => {
     (api.generateSegments as any).mockResolvedValue(undefined);
